Add logging support with training callsigns.
Unlicensed users can switch the logging mode to training and enter an existing trainings call.
There is a new contacted operator field which must be used for training and the system also logs the unlicensed call as the operator while using the training call of a licensed operator as the station callsign.
The trainer user account is associated to the QSOs, the plan is to allow them to edit and delete training QSOs as well, even though they did not create them.